SQS uses the bound S3 credentials when using shared bucket
