+++ test/api/support/cfAPINocks.js
@@ -1,5 +1,4 @@
 const nock = require('nock');
-const crypto = require('crypto');
 const factory = require('./factory');
 
 const url = 'https://api.example.com';
@@ -83,64 +82,6 @@ const mockCreateServiceKey = (body, resources) => {
   return n.reply(200, resources);
 };
 
-const mockSiteCreator = (
-  name,
-  region = 'us-gov-west-1',
-  bucket = 's3-bucket'
-) => {
-  const keyName = `${name}-key`;
-  const planName = 'basic-public';
-  const planGuid = 'plan-guid';
-  const bucketGuid = 'bucket-guid';
-  const accessKeyId = crypto.randomBytes(3).toString('hex');
-  const secretAccessKey = crypto.randomBytes(3).toString('hex');
-
-  const instanceRequestBody = { name, service_plan_guid: planGuid };
-  const keyRequestBody = { name, service_instance_guid: bucketGuid };
-
-  const planResponses = {
-    resources: [
-      factory.responses.service({ guid: planGuid }, { name: planName }),
-    ],
-  };
-  const bucketResponse = factory.responses.service({ guid: bucketGuid }, { name });
-  const keyResponse = factory.responses.service({}, {
-    name: keyName,
-    service_instance_guid: bucketGuid,
-    credentials: factory.responses.credentials({
-      access_key_id: accessKeyId,
-      secret_access_key: secretAccessKey,
-      region,
-      bucket,
-    }),
-  });
-
-  const buildResponses = {
-    resources: [
-      factory.responses.service({}, {
-        name,
-        service_instance_guid: bucketGuid,
-        credentials: factory.responses.credentials({
-          access_key_id: accessKeyId,
-          secret_access_key: secretAccessKey,
-          region,
-          bucket,
-        }),
-      }),
-    ],
-  };
-
-  const serviceCredentialsResponses = {
-    resources: [keyResponse],
-  };
-
-  mockFetchS3ServicePlanGUID(planResponses);
-  mockCreateS3ServiceInstance(instanceRequestBody, bucketResponse);
-  mockCreateServiceKey(keyRequestBody, keyResponse);
-  mockFetchServiceInstancesRequest(buildResponses);
-  mockFetchServiceInstanceCredentialsRequest('test-guid', serviceCredentialsResponses);
-};
-
 module.exports = {
   mockCreateS3ServiceInstance,
   mockCreateServiceKey,
@@ -150,5 +91,4 @@ module.exports = {
   mockFetchServiceKeyRequest,
   mockFetchServiceKeysRequest,
   mockFetchS3ServicePlanGUID,
-  mockSiteCreator,
 };
