+++ test/api/unit/services/SiteCreator.test.js
@@ -11,8 +11,6 @@ const TemplateResolver = require('../../../../api/services/TemplateResolver');
 const { Build, Site, User } = require('../../../../api/models');
 
 describe('SiteCreator', () => {
-  beforeEach(() => mockTokenRequest());
-
   afterEach(() => nock.cleanAll());
 
   describe('.createSite', () => {
@@ -62,8 +60,6 @@ describe('SiteCreator', () => {
             repository: crypto.randomBytes(3).toString('hex'),
           };
 
-          apiNocks.mockDefaultCredentials();
-
           factory.user().then((model) => {
             user = model;
             githubAPINocks.repo();
@@ -101,8 +97,6 @@ describe('SiteCreator', () => {
             repository: crypto.randomBytes(3).toString('hex'),
           };
 
-          apiNocks.mockDefaultCredentials();
-
           factory.user().then((model) => {
             user = model;
             githubAPINocks.repo();
@@ -134,8 +128,6 @@ describe('SiteCreator', () => {
           const siteParams = {
             repository: crypto.randomBytes(3).toString('hex'),
           };
-
-          apiNocks.mockDefaultCredentials();
 
           factory.user().then((model) => {
             user = model;
@@ -268,8 +260,6 @@ describe('SiteCreator', () => {
       let user;
 
       it('should create a new site record for the given repository and add the user', (done) => {
-        apiNocks.mockDefaultCredentials();
-
         factory.user().then((model) => {
           user = model;
           githubAPINocks.createRepoForOrg();
@@ -317,8 +307,6 @@ describe('SiteCreator', () => {
           owner: '18f',
           branch: 'not-master',
         };
-
-        apiNocks.mockDefaultCredentials();
 
         templateResolverStub.returns(fakeTemplate);
 
@@ -475,11 +463,62 @@ describe('SiteCreator', () => {
           sharedBucket: false,
         };
 
-        const name = siteParams.repository;
-        const s3ServiceName = `${siteParams.owner}-${siteParams.repository}`;
+        const name = `${siteParams.owner}-${siteParams.repository}`;
+        const keyName = `${name}-key`;
+        const planName = 'basic-public';
+        const planGuid = 'plan-guid';
+        const bucketGuid = 'bucket-guid';
+        const accessKeyId = crypto.randomBytes(3).toString('hex');
+        const secretAccessKey = crypto.randomBytes(3).toString('hex');
         const region = 'us-gov-other-1';
         const bucket = 'testing-bucket';
-        apiNocks.mockSiteCreator(name, region, bucket);
+
+        const instanceRequestBody = { name, service_plan_guid: planGuid };
+        const keyRequestBody = { name, service_instance_guid: bucketGuid };
+
+        const planResponses = {
+          resources: [
+            factory.responses.service({ guid: planGuid }, { name: planName }),
+          ],
+        };
+        const bucketResponse = factory.responses.service({ guid: bucketGuid }, { name });
+        const keyResponse = factory.responses.service({}, {
+          name: keyName,
+          service_instance_guid: bucketGuid,
+          credentials: factory.responses.credentials({
+            access_key_id: accessKeyId,
+            secret_access_key: secretAccessKey,
+            region,
+            bucket,
+          }),
+        });
+
+        const buildResponses = {
+          resources: [
+            factory.responses.service({}, {
+              name,
+              service_instance_guid: bucketGuid,
+              credentials: factory.responses.credentials({
+                access_key_id: accessKeyId,
+                secret_access_key: secretAccessKey,
+                region,
+                bucket,
+              }),
+            }),
+          ],
+        };
+
+        const serviceCredentialsResponses = {
+          resources: [keyResponse],
+        };
+
+        mockTokenRequest();
+        apiNocks.mockFetchS3ServicePlanGUID(planResponses);
+        apiNocks.mockCreateS3ServiceInstance(instanceRequestBody, bucketResponse);
+        apiNocks.mockCreateServiceKey(keyRequestBody, keyResponse);
+        apiNocks.mockFetchServiceInstancesRequest(buildResponses);
+        apiNocks.mockFetchServiceInstanceCredentialsRequest('test-guid', serviceCredentialsResponses);
+
         factory.user().then((model) => {
           user = model;
           githubAPINocks.repo();
@@ -504,7 +543,7 @@ describe('SiteCreator', () => {
               siteParams.owner,
               siteParams.repository,
               user,
-              s3ServiceName,
+              name,
               bucket,
               region
             );
